+++ engine/src/communication/CommunicationInterface/messageSender.cpp
@@ -1,4 +1,7 @@
 #include "messageSender.hpp"
+#include <algorithm>
+
+using namespace fmt::literals;
 
 namespace comm {
 
@@ -6,7 +9,7 @@ message_sender * message_sender::instance = nullptr;
 
 message_sender * message_sender::get_instance() {
 	if(instance == NULL) {
-		throw std::exception();
+		throw std::runtime_error("ERROR: message_sender::get_instance() had a NULL instance");
 	}
 	return instance;
 }
@@ -17,7 +20,7 @@ void message_sender::initialize_instance(std::shared_ptr<ral::cache::CacheMachin
 		int num_threads,
 		ucp_context_h context,
 		ucp_worker_h origin_node,
-		uint16_t ral_id,
+		int ral_id,
 		comm::blazing_protocol protocol){
 	
 	if(instance == NULL) {
@@ -32,7 +35,7 @@ message_sender::message_sender(std::shared_ptr<ral::cache::CacheMachine> output_
 		int num_threads,
 		ucp_context_h context,
 		ucp_worker_h origin,
-		uint16_t ral_id,
+		int ral_id,
 		comm::blazing_protocol protocol)
 		: pool{num_threads}, output_cache{output_cache}, input_cache{input_cache}, node_address_map{node_address_map}, protocol{protocol}, origin{origin}, ral_id{ral_id}
 {
@@ -59,21 +62,37 @@ void message_sender::run_polling() {
 		polling_started = true;
 
 		auto thread = std::thread([this]{
+		std::shared_ptr<spdlog::logger> comms_logger;
+		comms_logger = spdlog::get("output_comms");
+
 		cudaSetDevice(0);
 
 		while(true) {
 			std::vector<std::unique_ptr<ral::cache::CacheData> > cache_datas = output_cache->pull_all_cache_data();
 			for(auto & cache_data : cache_datas){
+
 				pool.push([cache_data{std::move(cache_data)},
 						node_address_map = node_address_map,
 						output_cache = output_cache,
 							protocol=this->protocol,
-							this](int /*thread_id*/) {
+							this,
+							comms_logger](int /*thread_id*/) {
 
 					auto * gpu_cache_data = static_cast<ral::cache::GPUCacheDataMetaData *>(cache_data.get());
 					auto data_and_metadata = gpu_cache_data->decacheWithMetaData();
 					auto & metadata = data_and_metadata.second;
 					auto & table = data_and_metadata.first;
+
+
+					auto destinations = metadata.get_values()[ral::cache::WORKER_IDS_METADATA_LABEL];
+					comms_logger->info("{ral_id}|{query_id}|{kernel_id}|{dest_ral_id}|{dest_ral_count}|{dest_cache_id}|{message_id}",
+						"ral_id"_a=ral_id,
+						"query_id"_a=metadata.get_values()[ral::cache::QUERY_ID_METADATA_LABEL],
+						"kernel_id"_a=metadata.get_values()[ral::cache::KERNEL_ID_METADATA_LABEL],
+						"dest_ral_id"_a=destinations, //false
+						"dest_ral_count"_a=std::count(destinations.begin(), destinations.end(), ',') + 1,
+						"dest_cache_id"_a=metadata.get_values()[ral::cache::CACHE_ID_METADATA_LABEL],
+						"message_id"_a=metadata.get_values()[ral::cache::MESSAGE_ID]);
 
 					std::vector<std::size_t> buffer_sizes;
 					std::vector<const char *> raw_buffers;
@@ -123,7 +142,12 @@ void message_sender::run_polling() {
 							transport->send(raw_buffers[i], buffer_sizes[i]);
 						}
 						transport->wait_until_complete();  // ensures that the message has been sent before returning the thread to the pool
-					} catch(const std::exception&) {
+					} catch(const std::exception & e) {
+						auto logger = spdlog::get("batch_logger");
+						if (logger){
+							logger->error("|||{info}|||||",
+									"info"_a="ERROR in message_sender::run_polling(). What: {}"_format(e.what()));
+						}
 						throw;
 					}
 			});