+++ engine/src/execution_graph/logic_controllers/CacheMachine.cpp
@@ -5,6 +5,7 @@
 #include <cudf/io/orc.hpp>
 #include "CalciteExpressionParsing.h"
 #include "communication/CommunicationData.h"
+#include <Util/StringUtil.h>
 #include <stdio.h>
 
 using namespace std::chrono_literals;
@@ -55,18 +56,18 @@ std::unique_ptr<CacheData> CacheData::downgradeCacheData(std::unique_ptr<CacheDa
 			return std::make_unique<CPUCacheData>(std::move(table));
 		
 		} else {
+			// want to get only cache directory where orc files should be saved
+			std::string orc_files_path = ral::communication::CommunicationData::getInstance().get_cache_directory();
 			if(logger != nullptr) {
 				logger->trace("{query_id}|{step}|{substep}|{info}||kernel_id|{kernel_id}|rows|{rows}",
 					"query_id"_a=(ctx ? std::to_string(ctx->getContextToken()) : ""),
 					"step"_a=(ctx ? std::to_string(ctx->getQueryStep()) : ""),
 					"substep"_a=(ctx ? std::to_string(ctx->getQuerySubstep()) : ""),
-					"info"_a="Downgraded CacheData to Disk cache",
+					"info"_a="Downgraded CacheData to Disk cache to path: " + orc_files_path,
 					"kernel_id"_a=id,
 					"rows"_a=table->num_rows());
 			}
 
-			// want to get only cache directory where orc files should be saved
-			std::string orc_files_path = ral::communication::CommunicationData::getInstance().get_cache_directory();
 			return std::make_unique<CacheDataLocalFile>(std::move(table), orc_files_path, (ctx ? std::to_string(ctx->getContextToken()) : "none"));
 		}
 	}	
@@ -86,6 +87,7 @@ size_t CacheDataLocalFile::sizeInBytes() const {
 }
 
 std::unique_ptr<ral::frame::BlazingTable> CacheDataLocalFile::decache() {
+
 	cudf::io::orc_reader_options read_opts = cudf::io::orc_reader_options::builder(cudf::io::source_info{this->filePath_});
 	auto result = cudf::io::read_orc(read_opts);
 
@@ -168,15 +170,35 @@ CacheDataLocalFile::CacheDataLocalFile(std::unique_ptr<ral::frame::BlazingTable>
 	this->size_in_bytes = table->sizeInBytes();
 	this->filePath_ = orc_files_path + "/.blazing-temp-" + ctx_token + "-" + randomString(64) + ".orc";
 
-	cudf::io::table_metadata metadata;
-	for(auto name : table->names()) {
-		metadata.column_names.emplace_back(name);
+	int attempts = 0;
+	int attempts_limit = 10;
+	while(attempts <= attempts_limit){
+		try {
+			cudf::io::table_metadata metadata;
+			for(auto name : table->names()) {
+				metadata.column_names.emplace_back(name);
+			}
+
+			cudf::io::orc_writer_options out_opts = cudf::io::orc_writer_options::builder(cudf::io::sink_info{this->filePath_}, table->view())
+				.metadata(&metadata);
+
+			cudf::io::write_orc(out_opts);
+		} catch (cudf::logic_error & err){
+			auto logger = spdlog::get("batch_logger");
+			if(logger != nullptr) {
+				logger->error("|||{info}||||rows|{rows}",
+					"info"_a="Failed to create CacheDataLocalFile in path: " + this->filePath_ + " attempt " + std::to_string(attempts),
+					"rows"_a=table->num_rows());
+			}	
+			attempts++;
+			if (attempts == attempts_limit){
+				throw;
+			}
+			std::this_thread::sleep_for (std::chrono::milliseconds(5 * attempts));
+		}
 	}
 
-	cudf::io::orc_writer_options out_opts = cudf::io::orc_writer_options::builder(cudf::io::sink_info{this->filePath_}, table->view())
-		.metadata(&metadata);
-
-	cudf::io::write_orc(out_opts);
+	
 }
 
 ConcatCacheData::ConcatCacheData(std::vector<std::unique_ptr<CacheData>> cache_datas, const std::vector<std::string>& col_names, const std::vector<cudf::data_type>& schema)
@@ -223,11 +245,12 @@ std::unique_ptr<GPUCacheDataMetaData> cast_cache_data_to_gpu_with_meta(std::uniq
 }
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-CacheMachine::CacheMachine(std::shared_ptr<Context> context, bool log_timeout): ctx(context), cache_id(CacheMachine::cache_count)
+CacheMachine::CacheMachine(std::shared_ptr<Context> context, std::string cache_machine_name, bool log_timeout): 
+	ctx(context), cache_id(CacheMachine::cache_count), cache_machine_name(cache_machine_name)
 {
 	CacheMachine::cache_count++;
 
-	waitingCache = std::make_unique<WaitingQueue <std::unique_ptr <message> > >(60000, log_timeout);
+	waitingCache = std::make_unique<WaitingQueue <std::unique_ptr <message> > >(cache_machine_name, 60000, log_timeout);
 	this->memory_resources.push_back( &blazing_device_memory_resource::getInstance() );
 	this->memory_resources.push_back( &blazing_host_memory_resource::getInstance() );
 	this->memory_resources.push_back( &blazing_disk_memory_resource::getInstance() );
@@ -263,13 +286,14 @@ std::int32_t CacheMachine::get_id() const { return (cache_id); }
 void CacheMachine::finish() {
 	this->waitingCache->finish();
 	if(logger != nullptr) {
-		logger->trace("{query_id}|{step}|{substep}|{info}|{duration}|cache_id|{cache_id}||",
+		logger->trace("{query_id}|{step}|{substep}|{info}|{duration}|cache_id|{cache_id}|cache_name|{cache_name}",
 									"query_id"_a=(ctx ? std::to_string(ctx->getContextToken()) : ""),
 									"step"_a=(ctx ? std::to_string(ctx->getQueryStep()) : ""),
 									"substep"_a=(ctx ? std::to_string(ctx->getQuerySubstep()) : ""),
 									"info"_a="CacheMachine finish()",
 									"duration"_a="",
-									"cache_id"_a=cache_id);
+									"cache_id"_a=cache_id,
+									"cache_name"_a=cache_machine_name);
 	}
 }
 
@@ -289,7 +313,7 @@ uint64_t CacheMachine::get_num_batches_added(){
 	return this->waitingCache->processed_parts();
 }
 
-bool CacheMachine::addHostFrameToCache(std::unique_ptr<ral::frame::BlazingHostTable> host_table, const std::string & message_id) {
+bool CacheMachine::addHostFrameToCache(std::unique_ptr<ral::frame::BlazingHostTable> host_table, std::string message_id) {
 
 	// we dont want to add empty tables to a cache, unless we have never added anything
 	if (!this->something_added || host_table->num_rows() > 0){
@@ -306,6 +330,11 @@ bool CacheMachine::addHostFrameToCache(std::unique_ptr<ral::frame::BlazingHostTa
 
 		num_rows_added += host_table->num_rows();
 		num_bytes_added += host_table->sizeInBytes();
+
+		if (message_id == ""){
+			message_id = this->cache_machine_name;
+		}
+
 		auto cache_data = std::make_unique<CPUCacheData>(std::move(host_table));
 		auto item =	std::make_unique<message>(std::move(cache_data), message_id);
 		this->waitingCache->put(std::move(item));
@@ -317,8 +346,8 @@ bool CacheMachine::addHostFrameToCache(std::unique_ptr<ral::frame::BlazingHostTa
 	return false;
 }
 
-void CacheMachine::put(size_t message_id, std::unique_ptr<ral::frame::BlazingTable> table) {
-	this->addToCache(std::move(table), std::to_string(message_id), true);
+void CacheMachine::put(size_t index, std::unique_ptr<ral::frame::BlazingTable> table) {
+	this->addToCache(std::move(table), this->cache_machine_name + "_" + std::to_string(index), true);
 }
 
 void CacheMachine::clear() {
@@ -339,11 +368,19 @@ std::vector<std::unique_ptr<ral::cache::CacheData> > CacheMachine::pull_all_cach
 	return std::move(new_messages);
 }
 
-std::vector<std::string> CacheMachine::get_all_message_ids() {
-	return this->waitingCache->get_all_message_ids();
+std::vector<size_t> CacheMachine::get_all_indexes() {
+	std::vector<std::string> message_ids = this->waitingCache->get_all_message_ids();
+	std::vector<size_t> indexes;
+	indexes.reserve(message_ids.size());
+	for (auto & message_id : message_ids){
+		std::string prefix = this->cache_machine_name + "_";
+		assert(StringUtil::beginsWith(message_id, prefix));
+		indexes.push_back(std::stoi(message_id.substr(prefix.size())));
+	}
+	return indexes;
 }
 
-bool CacheMachine::addCacheData(std::unique_ptr<ral::cache::CacheData> cache_data, const std::string & message_id, bool always_add){
+bool CacheMachine::addCacheData(std::unique_ptr<ral::cache::CacheData> cache_data, std::string message_id, bool always_add){
 
 	// we dont want to add empty tables to a cache, unless we have never added anything
 	if ((!this->something_added || cache_data->num_rows() > 0) || always_add){
@@ -357,6 +394,10 @@ bool CacheMachine::addCacheData(std::unique_ptr<ral::cache::CacheData> cache_dat
 			cacheIndex = 1;
 		} else {
 			cacheIndex = 2;
+		}
+
+		if (message_id == ""){
+			message_id = this->cache_machine_name;
 		}
 
 		if(cacheIndex == 0) {
@@ -414,7 +455,7 @@ bool CacheMachine::addCacheData(std::unique_ptr<ral::cache::CacheData> cache_dat
 	return false;
 }
 
-bool CacheMachine::addToCache(std::unique_ptr<ral::frame::BlazingTable> table, const std::string & message_id, bool always_add) {
+bool CacheMachine::addToCache(std::unique_ptr<ral::frame::BlazingTable> table, std::string message_id, bool always_add) {
 	// we dont want to add empty tables to a cache, unless we have never added anything
 	if (!this->something_added || table->num_rows() > 0 || always_add){
 		for (auto col_ind = 0; col_ind < table->num_columns(); col_ind++){
@@ -424,6 +465,11 @@ bool CacheMachine::addToCache(std::unique_ptr<ral::frame::BlazingTable> table, c
 			}
 
 		}
+
+		if (message_id == ""){
+			message_id = this->cache_machine_name;
+		}
+
 		num_rows_added += table->num_rows();
 		num_bytes_added += table->sizeInBytes();
 		size_t cacheIndex = 0;
@@ -507,12 +553,42 @@ void CacheMachine::wait_until_finished() {
 
 
 std::unique_ptr<ral::frame::BlazingTable> CacheMachine::get_or_wait(size_t index) {
-	std::unique_ptr<message> message_data = waitingCache->get_or_wait(std::to_string(index));
+	std::unique_ptr<message> message_data = waitingCache->get_or_wait(this->cache_machine_name + "_" + std::to_string(index));
 	if (message_data == nullptr) {
 		return nullptr;
 	}
+	if (logger != nullptr){
+		logger->trace("{query_id}|{step}|{substep}|{info}|{duration}|kernel_id|{kernel_id}|rows|{rows}",
+								"query_id"_a=(ctx ? std::to_string(ctx->getContextToken()) : ""),
+								"step"_a=(ctx ? std::to_string(ctx->getQueryStep()) : ""),
+								"substep"_a=(ctx ? std::to_string(ctx->getQuerySubstep()) : ""),
+								"info"_a="CacheMachine::get_or_wait pulling from cache ",
+								"duration"_a="",
+								"kernel_id"_a=message_data->get_message_id(),
+								"rows"_a=message_data->get_data().num_rows());
+	}		
 
 	std::unique_ptr<ral::frame::BlazingTable> output = message_data->get_data().decache();
+	return std::move(output);
+}
+
+std::unique_ptr<ral::cache::CacheData>  CacheMachine::get_or_wait_CacheData(size_t index) {
+	std::unique_ptr<message> message_data = waitingCache->get_or_wait(this->cache_machine_name + "_" + std::to_string(index));
+	if (message_data == nullptr) {
+		return nullptr;
+	}
+	if (logger != nullptr){
+		logger->trace("{query_id}|{step}|{substep}|{info}|{duration}|kernel_id|{kernel_id}|rows|{rows}",
+								"query_id"_a=(ctx ? std::to_string(ctx->getContextToken()) : ""),
+								"step"_a=(ctx ? std::to_string(ctx->getQueryStep()) : ""),
+								"substep"_a=(ctx ? std::to_string(ctx->getQuerySubstep()) : ""),
+								"info"_a="CacheMachine::get_or_wait pulling CacheData from cache ",
+								"duration"_a="",
+								"kernel_id"_a=message_data->get_message_id(),
+								"rows"_a=message_data->get_data().num_rows());
+	}		
+
+	std::unique_ptr<ral::cache::CacheData> output = message_data->release_data();
 	return std::move(output);
 }
 
@@ -639,12 +715,32 @@ size_t CacheMachine::downgradeCacheData() {
 	return bytes_downgraded;
 }
 
-ConcatenatingCacheMachine::ConcatenatingCacheMachine(std::shared_ptr<Context> context)
-	: CacheMachine(context) {}
+bool CacheMachine::has_messages_now(std::vector<std::string> messages){
+
+	std::vector<std::string> current_messages = this->waitingCache->get_all_message_ids();
+	for (auto & message : messages){
+		bool found = false;
+		for (auto & cur_message : current_messages){
+			if (message == cur_message)	{
+				found = true;
+				break;
+			}
+		}
+		if (!found){
+			return false;
+		}
+	}
+	return true;
+}
+
+	
+
+ConcatenatingCacheMachine::ConcatenatingCacheMachine(std::shared_ptr<Context> context, std::string cache_machine_name)
+	: CacheMachine(context, cache_machine_name) {}
 
 ConcatenatingCacheMachine::ConcatenatingCacheMachine(std::shared_ptr<Context> context,
-			std::size_t concat_cache_num_bytes, bool concat_all)
-	: CacheMachine(context), concat_cache_num_bytes(concat_cache_num_bytes), concat_all(concat_all) {
+			std::size_t concat_cache_num_bytes, bool concat_all, std::string cache_machine_name)
+	: CacheMachine(context, cache_machine_name), concat_cache_num_bytes(concat_cache_num_bytes), concat_all(concat_all) {
 
 	}
 
@@ -742,7 +838,7 @@ std::unique_ptr<ral::cache::CacheData> ConcatenatingCacheMachine::pullCacheData(
 	if (concat_all){
 		waitingCache->wait_until_finished();
 	} else {
-		waitingCache->wait_until_num_bytes(this->concat_cache_num_bytes);
+		waitingCache->wait_until_num_bytes(this->concat_cache_num_bytes);		
 	}
 
 	size_t total_bytes = 0;