+++ engine/src/execution_graph/logic_controllers/CacheMachine.h
@@ -41,6 +41,7 @@ using namespace fmt::literals;
 enum class CacheDataType { GPU, CPU, LOCAL_FILE, GPU_METADATA, IO_FILE, CONCATENATING };
 
 const std::string KERNEL_ID_METADATA_LABEL = "kernel_id"; /**< A message metadata field that indicates which kernel owns this message. */
+const std::string RAL_ID_METADATA_LABEL = "kernel_id"; /**< A message metadata field that indicates RAL ran this. */
 const std::string QUERY_ID_METADATA_LABEL = "query_id"; /**< A message metadata field that indicates which query owns this message. */
 const std::string CACHE_ID_METADATA_LABEL = "cache_id";  /**< A message metadata field that indicates what cache a message should be routed to. Can be empty string if
  																															and only if add_to_specific_cache == false. */
@@ -598,7 +599,8 @@ public:
 	/**
 	* Constructor
 	*/
-	WaitingQueue(int timeout = 60000, bool log_timeout = true) : finished{false}, timeout(timeout), log_timeout(log_timeout) {}
+	WaitingQueue(std::string queue_name, int timeout = 60000, bool log_timeout = true) : 
+		queue_name(queue_name), finished{false}, timeout(timeout), log_timeout(log_timeout) {}
 
 	/**
 	* Destructor
@@ -663,13 +665,31 @@ public:
 
 	void wait_for_count(int count){
 
+		CodeTimer blazing_timer;
 		std::unique_lock<std::mutex> lock(mutex_);
-		condition_variable_.wait(lock, [&, this] () {
-			if (count < this->processed){
-				throw std::runtime_error("WaitingQueue::wait_for_count encountered " + std::to_string(this->processed) + " when expecting " + std::to_string(count));
-			}
-			return count == this->processed;
-		});
+		while(!condition_variable_.wait_for(lock, timeout*1ms, [&, this] {
+				bool done_waiting = count == this->processed;
+				if (!done_waiting && blazing_timer.elapsed_time() > 59000 && this->log_timeout){
+					auto logger = spdlog::get("batch_logger");
+					if(logger != nullptr) {
+						logger->warn("|||{info}|{duration}||||",
+											"info"_a="WaitingQueue " + this->queue_name + " wait_for_count timed out. count = " + std::to_string(count) + " processed = " + std::to_string(this->processed),
+											"duration"_a=blazing_timer.elapsed_time());
+					}
+				}
+				if (count < this->processed){
+					throw std::runtime_error("WaitingQueue::wait_for_count " + this->queue_name + " encountered " + std::to_string(this->processed) + " when expecting " + std::to_string(count));
+				}
+				return done_waiting;
+			})){}
+
+
+		// condition_variable_.wait(lock, [&, this] () {
+		// 	if (count < this->processed){
+		// 		throw std::runtime_error("WaitingQueue::wait_for_count " + this->queue_name + " encountered " + std::to_string(this->processed) + " when expecting " + std::to_string(count));
+		// 	}
+		// 	return count == this->processed;
+		// });
 	}
 
 	/**
@@ -684,22 +704,22 @@ public:
 
 		CodeTimer blazing_timer;
 		std::unique_lock<std::mutex> lock(mutex_);
-		/*while(!condition_variable_.wait_for(lock, timeout*1ms, [&, this] {
+		while(!condition_variable_.wait_for(lock, timeout*1ms, [&, this] {
 				bool done_waiting = this->finished.load(std::memory_order_seq_cst) or !this->empty();
-				if (!done_waiting && blazing_timer.elapsed_time() > 59000){
+				if (!done_waiting && blazing_timer.elapsed_time() > 59000 && this->log_timeout){
 					auto logger = spdlog::get("batch_logger");
 					if(logger != nullptr) {
 						logger->warn("|||{info}|{duration}||||",
-											"info"_a="WaitingQueue pop_or_wait timed out",
+											"info"_a="WaitingQueue " + this->queue_name + " pop_or_wait timed out",
 											"duration"_a=blazing_timer.elapsed_time());
 					}
 				}
 				return done_waiting;
-			})){}*/
+			})){}
 
-		condition_variable_.wait(lock,[&, this] {
-				return this->finished.load(std::memory_order_seq_cst) or !this->empty();
-		});
+		// condition_variable_.wait(lock,[&, this] {
+		// 		return this->finished.load(std::memory_order_seq_cst) or !this->empty();
+		// });
 		if(this->message_queue_.size() == 0) {
 			return nullptr;
 		}
@@ -737,7 +757,7 @@ public:
 					auto logger = spdlog::get("batch_logger");
 					if(logger != nullptr) {
 						logger->warn("|||{info}|{duration}||||",
-											"info"_a="WaitingQueue wait_for_next timed out",
+											"info"_a="WaitingQueue " + this->queue_name + " wait_for_next timed out",
 											"duration"_a=blazing_timer.elapsed_time());
 					}
 				}
@@ -773,7 +793,7 @@ public:
 					auto logger = spdlog::get("batch_logger");
 					if(logger != nullptr) {
 					   logger->warn("|||{info}|{duration}||||",
-										   "info"_a="WaitingQueue wait_until_finished timed out",
+										   "info"_a="WaitingQueue " + this->queue_name + " wait_until_finished timed out",
  										   "duration"_a=blazing_timer.elapsed_time());
 					}
 				}
@@ -794,8 +814,8 @@ public:
 		std::unique_lock<std::mutex> lock(mutex_);
 		while(!condition_variable_.wait_for(lock, timeout*1ms, [&blazing_timer, num_bytes, this] {
 				bool done_waiting = this->finished.load(std::memory_order_seq_cst);
-				if (!done_waiting) {
-					size_t total_bytes = 0;
+				size_t total_bytes = 0;
+				if (!done_waiting) {					
 					for (auto & message : message_queue_){
 						total_bytes += message->get_data().sizeInBytes();
 					}
@@ -805,7 +825,7 @@ public:
 					auto logger = spdlog::get("batch_logger");
 					if(logger != nullptr) {
 						logger->warn("|||{info}|{duration}||||",
-											"info"_a="WaitingQueue wait_until_finished timed out",
+											"info"_a="WaitingQueue " + this->queue_name + " wait_until_num_bytes timed out num_bytes wanted: " + std::to_string(num_bytes) + " total_bytes: " + std::to_string(total_bytes),
 											"duration"_a=blazing_timer.elapsed_time());
 					}
 				}
@@ -853,7 +873,7 @@ public:
 					auto logger = spdlog::get("batch_logger");
 					if(logger != nullptr) {
 						logger->warn("|||{info}|{duration}|message_id|{message_id}||",
-											"info"_a="WaitingQueue get_or_wait timed out",
+											"info"_a="WaitingQueue " + this->queue_name + " get_or_wait timed out",
 											"duration"_a=blazing_timer.elapsed_time(),
 											"message_id"_a=message_id);
 					}
@@ -927,7 +947,7 @@ public:
 					auto logger = spdlog::get("batch_logger");
 					if(logger != nullptr) {
 						logger->warn("|||{info}|{duration}||||",
-											"info"_a="WaitingQueue get_all_or_wait timed out",
+											"info"_a="WaitingQueue " + this->queue_name + " get_all_or_wait timed out",
 											"duration"_a=blazing_timer.elapsed_time());
 					}
 				}
@@ -972,6 +992,7 @@ public:
 	void put_all(std::vector<message_ptr> messages){
 		std::unique_lock<std::mutex> lock(mutex_);
 		put_all_unsafe(std::move(messages));
+		processed += messages.size();
 		condition_variable_.notify_all();
 	}
 private:
@@ -999,6 +1020,7 @@ private:
 	int processed = 0; /**< Count of messages added to the WaitingQueue. */
 
 	int timeout; /**< timeout period in ms used by the wait_for to log that the condition_variable has been waiting for a long time. */
+	std::string queue_name;
 	bool log_timeout; /**< Whether or not to log when a timeout accurred. */
 };
 
@@ -1015,21 +1037,23 @@ std::unique_ptr<GPUCacheDataMetaData> cast_cache_data_to_gpu_with_meta(std::uniq
 */
 class CacheMachine {
 public:
-	CacheMachine(std::shared_ptr<Context> context, bool log_timeout = true);
+	CacheMachine(std::shared_ptr<Context> context, std::string cache_machine_name, bool log_timeout = true);
 
 	~CacheMachine();
 
-	virtual void put(size_t message_id, std::unique_ptr<ral::frame::BlazingTable> table);
+	virtual void put(size_t index, std::unique_ptr<ral::frame::BlazingTable> table);
 
 	virtual std::unique_ptr<ral::frame::BlazingTable> get_or_wait(size_t index);
 
+	virtual std::unique_ptr<ral::cache::CacheData> get_or_wait_CacheData(size_t index);
+
 	virtual void clear();
 
-	virtual bool addToCache(std::unique_ptr<ral::frame::BlazingTable> table, const std::string & message_id = "", bool always_add = false);
+	virtual bool addToCache(std::unique_ptr<ral::frame::BlazingTable> table, std::string message_id = "", bool always_add = false);
 
-	virtual bool addCacheData(std::unique_ptr<ral::cache::CacheData> cache_data, const std::string & message_id = "", bool always_add = false);
+	virtual bool addCacheData(std::unique_ptr<ral::cache::CacheData> cache_data, std::string message_id = "", bool always_add = false);
 
-	virtual bool addHostFrameToCache(std::unique_ptr<ral::frame::BlazingHostTable> table, const std::string & message_id = "");
+	virtual bool addHostFrameToCache(std::unique_ptr<ral::frame::BlazingHostTable> table, std::string message_id = "");
 
 	virtual void finish();
 
@@ -1055,6 +1079,8 @@ public:
 		return this->waitingCache->has_next_now();
 	}
 
+	bool has_messages_now(std::vector<std::string> messages);
+
 	std::size_t get_num_batches(){
 		return cache_count;
 	}
@@ -1068,7 +1094,7 @@ public:
 
 	virtual std::unique_ptr<ral::cache::CacheData> pullCacheData();
 
-	std::vector<std::string> get_all_message_ids();
+	std::vector<size_t> get_all_indexes();
 
 	void wait_for_count(int count){
 		return this->waitingCache->wait_for_count(count);
@@ -1094,6 +1120,7 @@ protected:
 	std::shared_ptr<spdlog::logger> logger;
 	std::shared_ptr<spdlog::logger> cache_events_logger;
 	const std::size_t cache_id;
+	std::string cache_machine_name;
 };
 
 /**
@@ -1104,7 +1131,7 @@ protected:
 class HostCacheMachine {
 public:
 	HostCacheMachine(std::shared_ptr<Context> context, const std::size_t id) : ctx(context), cache_id(id) {
-		waitingCache = std::make_unique<WaitingQueue <std::unique_ptr< message> > >();
+		waitingCache = std::make_unique<WaitingQueue <std::unique_ptr< message> > >("");
 		logger = spdlog::get("batch_logger");
 		something_added = false;
 
@@ -1196,10 +1223,10 @@ protected:
 */
 class ConcatenatingCacheMachine : public CacheMachine {
 public:
-	ConcatenatingCacheMachine(std::shared_ptr<Context> context);
+	ConcatenatingCacheMachine(std::shared_ptr<Context> context, std::string cache_machine_name);
 
 	ConcatenatingCacheMachine(std::shared_ptr<Context> context,
-			std::size_t concat_cache_num_bytes, bool concat_all);
+			std::size_t concat_cache_num_bytes, bool concat_all, std::string cache_machine_name);
 
 	~ConcatenatingCacheMachine() = default;