+++ engine/src/execution_graph/logic_controllers/taskflow/kernel.cpp
@@ -6,7 +6,7 @@ namespace ral {
 namespace cache {
 
 kernel::kernel(std::size_t kernel_id, std::string expr, std::shared_ptr<Context> context, kernel_type kernel_type_id)
-    : total_input_bytes{0}, expression{expr}, kernel_id(kernel_id), kernel_type_id{kernel_type_id}, context{context} {
+    : total_input_bytes_processed{0}, expression{expr}, kernel_id(kernel_id), kernel_type_id{kernel_type_id}, context{context} {
     parent_id_ = -1;
     has_limit_ = false;
     limit_rows_ = -1;
@@ -146,16 +146,23 @@ ral::execution::task_result kernel::process(std::vector<std::unique_ptr<ral::fra
     // if (this->has_limit_ && output->get_num_rows_added() >= this->limit_rows_) {
     //     return;
     // }
+
     size_t bytes = 0;
     for(auto & input : inputs){
         bytes += input->sizeInBytes();
     }
     auto result = do_process(std::move(inputs), output, stream, args);
     if(result.status == ral::execution::SUCCESS){
-        total_input_bytes += bytes; // increment this AFTER its been processed successfully
+        total_input_bytes_processed += bytes; // increment this AFTER its been processed successfully
+    }
+    else{
+        auto logger = spdlog::get("batch_logger");
+        if (logger){
+            logger->error("|||{info}|||||",
+                    "info"_a="ERROR in kernel::process trying to do do_process. Kernel name is: " + this->kernel_name() + " Kernel id is: " + std::to_string(this->kernel_id));
+        }
     }
     return std::move(result);
-
 }
 
 void kernel::add_task(size_t task_id){
@@ -179,8 +186,8 @@ std::size_t kernel::estimate_output_bytes(const std::vector<std::unique_ptr<ral:
     }
 
     // if we have already processed, then we can estimate based on previous inputs and outputs
-    if (total_input_bytes.load() > 0){
-        return (std::size_t)((double)input_bytes * ((double)this->output_.total_bytes_added()/(double)total_input_bytes.load()));
+    if (total_input_bytes_processed.load() > 0){
+        return (std::size_t)((double)input_bytes * ((double)this->output_.total_bytes_added()/(double)total_input_bytes_processed.load()));
     } else { // if we have not already any batches, then lets estimate that the output is the same as the input
         return input_bytes;
     }